# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Laravel Base Project API'
  description: 'Laravel Base Project API Documentation'
  version: 1.0.0
servers:
  -
    url: 'http://my-default-host.com'
    description: 'API Server'
paths:
...
  /api/v1/achievements:
    get:
      tags:
        - 'FocusFlow Achievements'
      summary: 'Get all achievements and user progress'
      operationId: 53a709ae9c5c6d18249dcb35447954c0
      responses:
        200:
          description: 'Achievements data'
          content:
            application/json:
              schema:
                type: array
                items:
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                    description:
                      type: string
                    icon:
                      type: string
                      nullable: true
                    condition_type:
                      type: string
                    condition_value:
                      type: integer
                    is_unlocked:
                      type: boolean
                    unlocked_at:
                      type: string
                      format: date-time
                      nullable: true
                  type: object
      security:
        -
          sanctum: []
  '/api/v1/achievements/{id}':
    get:
      tags:
        - 'FocusFlow Achievements'
      summary: 'Get a specific achievement'
      operationId: 26103882964c79164cbc2f1d7b622a89
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Achievement details'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Achievement'
      security:
        -
          sanctum: []
...
  /api/v1/goals:
    get:
      tags:
        - 'FocusFlow Goals'
      summary: 'Get all goals'
      operationId: 118d00b61f67ab24537caed7cd21b42c
      parameters:
        -
          name: type
          in: query
          required: false
          schema:
            type: string
            enum:
              - daily
              - weekly
              - monthly
              - yearly
        -
          name: completed
          in: query
          required: false
          schema:
            type: boolean
      responses:
        200:
          description: 'List of goals'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Goal'
      security:
        -
          sanctum: []
    post:
      tags:
        - 'FocusFlow Goals'
      summary: 'Create a new goal'
      operationId: 8c696ef1752aa200023b373435d79320
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - title
                - type
                - start_date
              properties:
                title:
                  type: string
                  example: 'Learn PHP'
                description:
                  type: string
                  nullable: true
                type:
                  type: string
                  enum:
                    - daily
                    - weekly
                    - monthly
                    - yearly
                start_date:
                  type: string
                  format: date
                  example: '2025-01-01'
                target_date:
                  type: string
                  format: date
                  nullable: true
                notes:
                  type: string
                  nullable: true
                sub_goals:
                  type: array
                  items:
                    properties:
                      title:
                        type: string
                        example: Basics
                    type: object
              type: object
      responses:
        201:
          description: 'Goal created'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
      security:
        -
          sanctum: []
  '/api/v1/goals/{id}':
    get:
      tags:
        - 'FocusFlow Goals'
      summary: 'Get a specific goal'
      operationId: 6383c722bd8bb549a8d44d5d6f2e8392
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Goal details'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
      security:
        -
          sanctum: []
    put:
      tags:
        - 'FocusFlow Goals'
      summary: 'Update a goal'
      operationId: cee9c5638ce987bfced6946b7bb2d98f
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                title:
                  type: string
                description:
                  type: string
                  nullable: true
                type:
                  type: string
                  enum:
                    - daily
                    - weekly
                    - monthly
                    - yearly
                target_date:
                  type: string
                  format: date
                  nullable: true
                notes:
                  type: string
                  nullable: true
                progress:
                  type: integer
                  maximum: 100
                  minimum: 0
                sub_goals:
                  type: array
                  items:
                    properties:
                      title:
                        type: string
                      completed:
                        type: boolean
                    type: object
              type: object
      responses:
        200:
          description: 'Goal updated'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
      security:
        -
          sanctum: []
    delete:
      tags:
        - 'FocusFlow Goals'
      summary: 'Delete a goal'
      operationId: f7b1477b3da47afca837913d03a49bc8
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Goal deleted'
      security:
        -
          sanctum: []
  '/api/v1/goals/{id}/complete':
    post:
      tags:
        - 'FocusFlow Goals'
      summary: 'Complete a goal'
      operationId: 81207e9aad94949a9ff9c8091d24c1cd
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Goal completed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
      security:
        -
          sanctum: []
...
components:
  schemas:
    Achievement:
      title: Achievement
      description: 'Achievement model'
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
        description:
          type: string
        icon:
          type: string
          nullable: true
        condition_type:
          type: string
        condition_value:
          type: integer
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      type: object
    Goal:
      title: 'Goal Model'
      description: 'FocusFlow Goal model'
      required:
        - title
        - user_id
        - type
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        user_id:
          type: integer
          example: 1
        title:
          type: string
          example: 'Learn Laravel'
        description:
          type: string
          example: 'Master the Laravel framework'
          nullable: true
        type:
          type: string
          enum:
            - daily
            - weekly
            - monthly
            - yearly
          example: monthly
        progress:
          description: 'Progress percentage 0-100'
          type: integer
          example: 50
        start_date:
          type: string
          format: date
          example: '2025-01-01'
        target_date:
          type: string
          format: date
          example: '2025-01-31'
        completed:
          type: boolean
          example: false
        notes:
          type: string
          nullable: true
        sub_goals:
          type: array
          items:
            $ref: '#/components/schemas/GoalSubGoal'
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      type: object
    GoalSubGoal:
      title: 'GoalSubGoal Model'
      description: 'FocusFlow Goal Subgoal model'
      required:
        - title
        - goal_id
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        goal_id:
          type: integer
          example: 1
        title:
          type: string
          example: 'Check emails'
        completed:
          type: boolean
          example: false
        order:
          type: integer
          example: 0
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      type: object
...
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Laravel Sanctum token authentication. Enter token in format (Bearer <token>)'
...